fix(review): fix invariant violations, a11y, structure, and minor cleanup
C-1: call rebuildTrackIdxMap() after tracks[] field mutations in orphans.js _relocateOrphan()
C-2: remove dead rebuildTrackIdxMap import from watchfolder.js; update header comment
M-1: add labelledby property + connectedCallback warning to lf-modal; fix focus trap to
     listen on host so slotted light-DOM keydown events are captured; add idempotency guard
     on close(); migrate FOCUSABLE_SEL inline to break feature-module import
M-3: extract updatePlActionBar() + updateBreadcrumb() to renderer-playlist-bar.js (122
     lines); renderer-grids.js reduced 819 to 706 lines; re-export chain preserved
N-1: delete unused trapFocusIn export from a11y.js (zero import sites)
N-2: remove non-standard :host-context() blocks from lf-toast-stack.js
N-3: replace 5 .onclick= assignments with addEventListener in views.js
N-4: replace console.error with console.warn in cfgsave.js, radio.js, watchfolder.js

All 439 tests pass. Final review gate: APPROVE.

+++ frontend/src/oscPremium.js
@@ -49,7 +49,7 @@ export function createPremiumOscilloscope(canvas, analyser) {
   // assumes analyser.fftSize >= SAMPLE_CAP (2048 >= 128 guaranteed by eq.js initEQ)
   const sampleCount = Math.min(SAMPLE_CAP, analyser.fftSize);
   const data     = new Uint8Array(sampleCount);
-  const prev     = new Uint8Array(sampleCount);   prev.fill(128);
+  const prev     = new Float32Array(sampleCount); prev.fill(128);
   const smoothed = new Float32Array(sampleCount); smoothed.fill(128);
 
   let dpr = 1;
